+++ python/cudf/io/csv.py
@@ -18,7 +18,15 @@ def _wrap_string(text):
         return ffi.new("char[]", text.encode())
 
 
-def read_csv(filepath, lineterminator='\n',
+def is_file_like(obj):
+    if not (hasattr(obj, 'read') or hasattr(obj, 'write')):
+        return False
+    if not hasattr(obj, "__iter__"):
+        return False
+    return True
+
+
+def read_csv(filepath_or_buffer, lineterminator='\n',
              quotechar='"', quoting=True, doublequote=True,
              delimiter=',', sep=None, delim_whitespace=False,
              skipinitialspace=False, names=None, dtype=None,
@@ -29,8 +37,8 @@ def read_csv(filepath, lineterminator='\n',
 
     Parameters
     ----------
-    filepath : str
-        Path of file to be read.
+    filepath_or_buffer : str
+        Path of file to be read or a file-like object containing the file.
     delimiter : char, default ','
         Delimiter to be used.
     delim_whitespace : bool, default False
@@ -98,8 +106,23 @@ def read_csv(filepath, lineterminator='\n',
     csv_reader = ffi.new('csv_read_arg*')
 
     # Populate csv_reader struct
-    file_path = _wrap_string(filepath)
-    csv_reader.file_path = file_path
+    if is_file_like(filepath_or_buffer):
+        buffer = filepath_or_buffer.read()
+        # check if StringIO is used
+        if hasattr(buffer, 'encode'):
+            buffer_as_bytes = buffer.encode()
+        else:
+            buffer_as_bytes = buffer
+        buffer_data_holder = ffi.new("char[]", buffer_as_bytes)
+
+        csv_reader.input_file.type = libgdf.BUFFER
+        csv_reader.input_file.buffer.data = buffer_data_holder
+        csv_reader.input_file.buffer.size = len(buffer_as_bytes)
+    else:
+        file_path = _wrap_string(filepath_or_buffer)
+
+        csv_reader.input_file.type = libgdf.FILE_PATH
+        csv_reader.input_file.path = file_path
 
     arr_names = []
     arr_dtypes = []
@@ -165,7 +188,7 @@ def read_csv(filepath, lineterminator='\n',
     return df
 
 
-def read_csv_strings(filepath, lineterminator='\n',
+def read_csv_strings(filepath_or_buffer, lineterminator='\n',
                      quotechar='"', quoting=True, doublequote=True,
                      delimiter=',', sep=None, delim_whitespace=False,
                      skipinitialspace=False, names=None, dtype=None,
@@ -235,8 +258,23 @@ def read_csv_strings(filepath, lineterminator='\n',
     csv_reader = ffi.new('csv_read_arg*')
 
     # Populate csv_reader struct
-    file_path = _wrap_string(filepath)
-    csv_reader.file_path = file_path
+    if is_file_like(filepath_or_buffer):
+        buffer = filepath_or_buffer.read()
+        # check if StringIO is used
+        if hasattr(buffer, 'encode'):
+            buffer_as_bytes = buffer.encode()
+        else:
+            buffer_as_bytes = buffer
+        buffer_data_holder = ffi.new("char[]", buffer_as_bytes)
+
+        csv_reader.input_file.type = libgdf.BUFFER
+        csv_reader.input_file.buffer.data = buffer_data_holder
+        csv_reader.input_file.buffer.size = len(buffer_as_bytes)
+    else:
+        file_path = _wrap_string(filepath_or_buffer)
+
+        csv_reader.input_file.type = libgdf.FILE_PATH
+        csv_reader.input_file.path = file_path
 
     arr_names = []
     arr_dtypes = []