+++ python/cudf/tests/test_csvreader.py
@@ -6,6 +6,9 @@ from collections import OrderedDict
 import numpy as np
 import pandas as pd
 
+from io import StringIO
+from io import BytesIO
+
 from cudf import read_csv
 from cudf.io.csv import read_csv_strings
 import cudf
@@ -264,3 +267,59 @@ def test_csv_reader_thousands(tmpdir):
     np.testing.assert_allclose(f64_ref, df['float64'])
     np.testing.assert_allclose(int32_ref, df['int32'])
     np.testing.assert_allclose(int64_ref, df['int64'])
+
+
+def test_csv_reader_buffer(tmpdir):
+
+    names = dtypes = ["float32", "int32", "date"]
+    lines = [','.join(names),
+             "1234.5, 1234567, 11/22/1995",
+             "12345.6, 12345, 1/2/2002"]
+
+    buffer = '\n'.join(lines) + '\n'
+
+    f32_ref = [1234.5, 12345.6]
+    int32_ref = [1234567, 12345]
+
+    df_str = read_csv(StringIO(buffer),
+                      names=names, dtype=dtypes, skiprows=1)
+    np.testing.assert_allclose(f32_ref, df_str['float32'])
+    np.testing.assert_allclose(int32_ref, df_str['int32'])
+    assert("1995-11-22T00:00:00.000" == str(df_str['date'][0]))
+    assert("2002-01-02T00:00:00.000" == str(df_str['date'][1]))
+
+    df_bytes = read_csv(BytesIO(str.encode(buffer)),
+                        names=names, dtype=dtypes, skiprows=1)
+    np.testing.assert_allclose(f32_ref, df_bytes['float32'])
+    np.testing.assert_allclose(int32_ref, df_bytes['int32'])
+    assert("1995-11-22T00:00:00.000" == str(df_bytes['date'][0]))
+    assert("2002-01-02T00:00:00.000" == str(df_bytes['date'][1]))
+
+
+def test_csv_reader_buffer_strings(tmpdir):
+
+    names = ['text', 'int']
+    dtypes = ['str', 'int']
+    lines = [','.join(names), 'a,0', 'b,0', 'c,0', 'd,0']
+
+    buffer = '\n'.join(lines) + '\n'
+
+    cols_str = read_csv_strings(StringIO(buffer),
+                                names=names, dtype=dtypes, skiprows=1)
+    assert(len(cols_str) == 2)
+    assert(type(cols_str[0]) == nvstrings.nvstrings)
+    assert(type(cols_str[1]) == cudf.Series)
+    assert(cols_str[0].sublist([0]).to_host()[0] == 'a')
+    assert(cols_str[0].sublist([1]).to_host()[0] == 'b')
+    assert(cols_str[0].sublist([2]).to_host()[0] == 'c')
+    assert(cols_str[0].sublist([3]).to_host()[0] == 'd')
+
+    cols_bytes = read_csv_strings(BytesIO(str.encode(buffer)),
+                                  names=names, dtype=dtypes, skiprows=1)
+    assert(len(cols_bytes) == 2)
+    assert(type(cols_bytes[0]) == nvstrings.nvstrings)
+    assert(type(cols_bytes[1]) == cudf.Series)
+    assert(cols_bytes[0].sublist([0]).to_host()[0] == 'a')
+    assert(cols_bytes[0].sublist([1]).to_host()[0] == 'b')
+    assert(cols_bytes[0].sublist([2]).to_host()[0] == 'c')
+    assert(cols_bytes[0].sublist([3]).to_host()[0] == 'd')